# Add generic AstVisitor