--- a/src/Esprima/Ast/AccessorProperty.cs
+++ b/src/Esprima/Ast/AccessorProperty.cs
@@ -28,7 +28,7 @@ public sealed class AccessorProperty : ClassProperty
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNextNullableAt2(Decorators, Key, Value);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitAccessorProperty(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitAccessorProperty(this);
 
     public AccessorProperty UpdateWith(Expression key, Expression? value, in NodeList<Decorator> decorators)
     {