--- a/src/Esprima/Ast/ArrayPattern.cs
+++ b/src/Esprima/Ast/ArrayPattern.cs
@@ -19,7 +19,7 @@ public sealed class ArrayPattern : BindingPattern
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNextNullable(Elements);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitArrayPattern(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitArrayPattern(this);
 
     public ArrayPattern UpdateWith(in NodeList<Node?> elements)
     {