--- a/src/Esprima/Ast/BinaryExpression.cs
+++ b/src/Esprima/Ast/BinaryExpression.cs
@@ -125,7 +125,7 @@ public class BinaryExpression : Expression
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Left, Right);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitBinaryExpression(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitBinaryExpression(this);
 
     protected virtual BinaryExpression Rewrite(Expression left, Expression right)
     {