--- a/src/Esprima/Ast/BreakStatement.cs
+++ b/src/Esprima/Ast/BreakStatement.cs
@@ -14,7 +14,7 @@ public sealed class BreakStatement : Statement
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNextNullable(Label);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitBreakStatement(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitBreakStatement(this);
 
     public BreakStatement UpdateWith(Identifier? label)
     {