--- a/src/Esprima/Ast/ClassBody.cs
+++ b/src/Esprima/Ast/ClassBody.cs
@@ -19,7 +19,7 @@ public sealed class ClassBody : Node
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Body);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitClassBody(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitClassBody(this);
 
     public ClassBody UpdateWith(in NodeList<ClassElement> body)
     {