--- a/src/Esprima/Ast/EmptyStatement.cs
+++ b/src/Esprima/Ast/EmptyStatement.cs
@@ -10,5 +10,5 @@ public sealed class EmptyStatement : Statement
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => null;
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitEmptyStatement(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitEmptyStatement(this);
 }