--- a/src/Esprima/Ast/ExportDefaultDeclaration.cs
+++ b/src/Esprima/Ast/ExportDefaultDeclaration.cs
@@ -17,7 +17,7 @@ public sealed class ExportDefaultDeclaration : ExportDeclaration
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Declaration);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitExportDefaultDeclaration(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitExportDefaultDeclaration(this);
 
     public ExportDefaultDeclaration UpdateWith(StatementListItem declaration)
     {