--- a/src/Esprima/Ast/ExpressionStatement.cs
+++ b/src/Esprima/Ast/ExpressionStatement.cs
@@ -14,7 +14,7 @@ public class ExpressionStatement : Statement
 
     internal sealed override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Expression);
 
-    protected internal sealed override object? Accept(AstVisitor visitor) => visitor.VisitExpressionStatement(this);
+    protected internal sealed override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitExpressionStatement(this);
 
     protected virtual ExpressionStatement Rewrite(Expression expression)
     {