--- a/src/Esprima/Ast/ForInStatement.cs
+++ b/src/Esprima/Ast/ForInStatement.cs
@@ -24,7 +24,7 @@ public sealed class ForInStatement : Statement
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Left, Right, Body);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitForInStatement(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitForInStatement(this);
 
     public ForInStatement UpdateWith(Node left, Expression right, Statement body)
     {