--- a/src/Esprima/Ast/ForOfStatement.cs
+++ b/src/Esprima/Ast/ForOfStatement.cs
@@ -27,7 +27,7 @@ public sealed class ForOfStatement : Statement
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Left, Right, Body);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitForOfStatement(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitForOfStatement(this);
 
     public ForOfStatement UpdateWith(Node left, Expression right, Statement body)
     {