--- a/src/Esprima/Ast/FunctionDeclaration.cs
+++ b/src/Esprima/Ast/FunctionDeclaration.cs
@@ -40,7 +40,7 @@ public sealed class FunctionDeclaration : Declaration, IFunction
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNextNullableAt0(Id, Params, Body);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitFunctionDeclaration(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitFunctionDeclaration(this);
 
     public FunctionDeclaration UpdateWith(Identifier? id, in NodeList<Node> parameters, BlockStatement body)
     {