--- a/src/Esprima/Ast/Identifier.cs
+++ b/src/Esprima/Ast/Identifier.cs
@@ -14,5 +14,5 @@ public sealed class Identifier : Expression
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => null;
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitIdentifier(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitIdentifier(this);
 }