--- a/src/Esprima/Ast/ImportAttribute.cs
+++ b/src/Esprima/Ast/ImportAttribute.cs
@@ -19,7 +19,7 @@ public sealed class ImportAttribute : Node
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Key, Value);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitImportAttribute(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitImportAttribute(this);
 
     public ImportAttribute UpdateWith(Expression key, Literal value)
     {