--- a/src/Esprima/Ast/LabeledStatement.cs
+++ b/src/Esprima/Ast/LabeledStatement.cs
@@ -17,7 +17,7 @@ public sealed class LabeledStatement : Statement
 
     internal override Node? NextChildNode(ref ChildNodes.Enumerator enumerator) => enumerator.MoveNext(Label, Body);
 
-    protected internal override object? Accept(AstVisitor visitor) => visitor.VisitLabeledStatement(this);
+    protected internal override T Accept<T>(AstVisitor<T> visitor) => visitor.VisitLabeledStatement(this);
 
     public LabeledStatement UpdateWith(Identifier label, Statement body)
     {