… src/Esprima/Ast/Literal.cs
… src/Esprima/Ast/MemberExpression.cs
… src/Esprima/Ast/MetaProperty.cs
… src/Esprima/Ast/MethodDefinition.cs
… src/Esprima/Ast/NewExpression.cs
… src/Esprima/Ast/Node.cs
… src/Esprima/Ast/ObjectExpression.cs
… src/Esprima/Ast/ObjectPattern.cs
… src/Esprima/Ast/PrivateIdentifier.cs
… src/Esprima/Ast/Program.cs
… src/Esprima/Ast/Property.cs
… src/Esprima/Ast/PropertyDefinition.cs
… src/Esprima/Ast/RestElement.cs
… src/Esprima/Ast/ReturnStatement.cs
… src/Esprima/Ast/SequenceExpression.cs
… src/Esprima/Ast/SpreadElement.cs
… src/Esprima/Ast/StaticBlock.cs
… src/Esprima/Ast/Super.cs
… src/Esprima/Ast/SwitchCase.cs
… src/Esprima/Ast/SwitchStatement.cs
… src/Esprima/Ast/TaggedTemplateExpression.cs
… src/Esprima/Ast/TemplateElement.cs
… src/Esprima/Ast/TemplateLiteral.cs
… src/Esprima/Ast/ThisExpression.cs
… src/Esprima/Ast/ThrowStatement.cs
… src/Esprima/Ast/TryStatement.cs
… src/Esprima/Ast/UnaryExpression.cs
… src/Esprima/Ast/VariableDeclaration.cs
… src/Esprima/Ast/VariableDeclarator.cs
… src/Esprima/Ast/WhileStatement.cs