This PR adds a generic AstVisitor (`AstVisitor<T>`).

This allows value types to be returned without boxing, for example a `double` to make a simple calculator:

https://github.com/sebastienros/esprima-dotnet/blob/977356b3aee00ba5d884d2a9eec8c66635bf9ee2/test/Esprima.Tests/AstVisitorTests.cs#L105-L141

 `AstVisitor` extends `AstVisitor<object?>` so it's still possible to use inherit `AstVisitor` without making a breaking change.