@adams85 2023-01-19
Although I've only got to take a quick look at the proposal yet, I have doubts because making virtual methods generic doesn't usually come without a cost:

## Esprima.Benchmark.VisitorBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old (.NET 6.0) |Visit|10.34 ms|4.592 ms|16 B|
| **New** (.NET 6.0)  |	| **13.76 ms (+33%)** | **1.815 ms** | **16 B (0%)** |
| Old (.NET Framework 4.7.2) |Visit|10.00 ms|2.900 ms|0 B|
| **New** (.NET Framework 4.7.2) |	| **13.76 ms (+38%)** | **3.110 ms** | **0 B** |
| Old (.NET 6.0)  |Rewrite|20.17 ms|5.604 ms|33 B|
| **New** (.NET 6.0)   |	| **24.44 ms (+21%)** | **7.346 ms** | **33 B (0%)** |
| Old (.NET Framework 4.7.2) |Rewrite|27.27 ms|4.751 ms|0 B|
| **New** (.NET Framework 4.7.2) |	| **30.70 ms (+13%)** | **1.566 ms** | **0 B** |

I like the idea in general but don't like the idea of taking the perf. hit in cases where reference return values are enough (or returns are not needed at all). And to me, this seems the (much?) more frequent case. I can't really see too many use cases which need value type returns. But maybe it's just my ignorance as e.g. Roslyn probably doesn't provide a non-generic and a generic visitor implementation with [parallel non-generic and generic `Accept` methods](https://github.com/dotnet/roslyn/blob/main/src/Compilers/CSharp/Portable/Syntax/CSharpSyntaxNode.cs#L117-L119) by accident.

Such a parallel solution would solve the perf. issue but would add to our maintenance burden significantly. Thinking of which, I don't really like this idea either...

Adding a generic `Accept` without the corresponding generic visitors might be something to consider though. This way who really need value type returns could implement the generic visitor for themselves.

But it's also worth noting that there may be other viable ways to handle this problem, for example object pooling.

These are my first thoughts. @lahma @jogibear9988 How about you?