@GerardSmit 2023-01-19
Huh, I didn't know that it would have this much of an impact on the performance. Good catch. I always thought that a generic method would "magically" create a new method with the generic types and having the same(ish) result.

I tried to change to make the final class sealed, and change the abstract class to an interface but both had the same performance hit.

I agree that the performance hit is too much for this to be merged, and having to maintain two methods (on every node) for both visitors isn't ideal either.

Alternatively we could create a source generator which
1. Creates a `AstVisitor<T>` class based on the `AstVisitor` class.
2. Adds `T Accept<T>(AstVisitor<T> visitor)` to every node.

But this is also more code to maintain which brings the question if it's worth it.