@sebastienros 2023-01-19
What if there is a generic version and a non generic one that inherits from the `Visitor<object>` ?
Would be non-breaking, but would the perf of the object one be the same?