@lahma 2023-01-20
Thank you @adams85 for reviewing and checking the perf impact! I think we need to be very careful in this regard as a lot of the esprima 3 (vnext) effort has been around performance (and of course features and fixes). I'd say @adams85 is the best person to give feedback on design as he's been the driving for around AST visitors (thank you for that!).

I'm trying  to understand the use case here. If it's about crafting an interpreter, maybe that should be a separate part - just like [Jint](https://github.com/sebastienros/jint) is built on top of the esprima AST and does the JavaScript interpretation. So in that way there's no need to "transform" return values.