@jogibear9988 2023-01-20
for me the perf hit does not matter, cause I only use esprima.net to parse and rewrite javascript files once during application startup.
maybe we could generate a copy of the Visitor, a generic one and a non generic one? Maybe the copy could be generated via Source Generators? Maybe add an Additional AcceptGeneric Method to each node?