@adams85 2023-01-20
Thank you all, guys, for sharing your opinions. They certainly helped my thought process.

> I'm trying to understand the use case here. If it's about crafting an interpreter, maybe that should be a separate part - just like [Jint](https://github.com/sebastienros/jint) is built on top of the esprima AST and does the JavaScript interpretation. So in that way there's no need to "transform" return values.

I can imagine a use case where you just want to evaluate some simple JS expressions (maybe only a very limited subset of what the language allows). I did [something similar in my PO parser library](https://github.com/adams85/po/blob/8b3c0ac570b03a43e45794731e9c2df0fb226725/source/Karambolo.PO/PluralExpression/PluralExpressionCompiler.cs). (Though that one doesn't evaluate the expression but parses it into a compiled lambda expression.) Anyway, the point is that such simple transformations wouldn't really require more than a visitor. And when you evaluate e.g. numeric expressions, it'd be nice to be able to return a value type to avoid boxing.

> Alternatively we could create a source generator [...]

> maybe we could generate a copy of the Visitor, a generic one and a non generic one? Maybe the copy could be generated via Source Generators? Maybe add an Additional AcceptGeneric Method to each node?

This is a nice idea, definitely worth exploring! I haven't had the time to delve into it but I suspect that a bunch of visitor-related boilerplate code could be generated. I think of stuff like
* AST methods `Node.Accept`, `Node.UpdateWith`, `Node.GetChildNodes` (this one is already generated but could be improved/done as part of a more general generation process) and
* general visitors `AstVisitor`, `AstRewriter`, `AstVisitorEventSource`.

There are a few special cases here and there but the whole thing follows a pretty solid pattern. We'd need to annotate our AST nodes to let the generator know which are the child properties, what is the default order of their visitation and probably a few other pieces of relevant information and we could generate about 90% of what I enumerated above. Handcrafting the remaining bits wouldn't be a big deal then.

In that case adding a generic version of `Accept` and visitors shouldn't be too much work either. As a matter of fact, this is the only viable way of adding this to the library which I can see now as I came to the conclusion that we'd better not take the performance hit. Which means the non-generic version must be kept as is, that is, the generic one could only be done separately.

The main problem is that implementing such a generator is quite a task... But I'm sure it'd be pretty damn exciting as well. So I may be not able to resist to give it a try. :) Of course, if anyone else has intentions like this, just let me know.