@sebastienros 2023-01-20
@adams85 
NB: I am also learning that you created a PO file parser (we did too in Orchard Core), and finally that you should start using [Parlot](https://github.com/sebastienros/parlot).