@adams85 2023-01-21
> I am also learning that you created a PO file parser (we did too in Orchard Core)

I know, took the idea of using PO for localization from Orchard (from the "original", not the Core version). 😄

> you should start using [Parlot](https://github.com/sebastienros/parlot).

Thanks for the tip. Looks like a nice lib, I'll take a closer look!