@adams85 2023-02-25
@GerardSmit 

> Alternatively we could create a source generator which
> 
>     Creates a AstVisitor<T> class based on the AstVisitor class.
>     Adds T Accept<T>(AstVisitor<T> visitor) to every node.

FYI, this is possible now: https://github.com/sebastienros/esprima-dotnet/pull/372

Do you feel like rewriting your PR based on this maybe?